+++ src/forall/debug/exceptions/InvalidArgumentException.php
@@ -1,0 +1,17 @@
+<?php
+
+/**
+ * @package forall.debug
+ * @author Avaq <aldwin.vlasblom@gmail.com>
+ */
+namespace forall\debug\exceptions;
+
+/**
+ * The invalid argument exception class.
+ */
+class InvalidArgumentException extends Exception
+{
+  
+  
+  
+}